+++ crates/header-translator/src/rust_type.rs
@@ -2301,13 +2301,21 @@ impl Ty {
         })
     }
 
-    fn fn_contains_bool_argument(&self) -> bool {
+    fn fn_contains_bool(&self) -> bool {
         if let Self::Pointer { pointee, .. } = self {
-            if let Self::Pointee(PointeeTy::Fn { arguments, .. }) = &**pointee {
+            if let Self::Pointee(PointeeTy::Fn {
+                arguments,
+                result_type,
+                ..
+            }) = &**pointee
+            {
                 if arguments
                     .iter()
                     .any(|arg| matches!(arg, Self::Primitive(Primitive::C99Bool)))
                 {
+                    return true;
+                }
+                if matches!(**result_type, Self::Primitive(Primitive::C99Bool)) {
                     return true;
                 }
             }
@@ -2321,7 +2329,7 @@ impl Ty {
             Self::Primitive(Primitive::Long) => write!(f, "Encoding::C_LONG"),
             Self::Primitive(Primitive::ULong) => write!(f, "Encoding::C_ULONG"),
             // TODO: Make all function pointers be encode, regardless of arguments
-            Self::TypeDef { to, .. } if to.fn_contains_bool_argument() => {
+            Self::TypeDef { to, .. } if to.fn_contains_bool() => {
                 write!(f, "Encoding::Pointer(&Encoding::Unknown)")
             }
             _ => write!(f, "<{}>::ENCODING", self.record()),